# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarRental API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rentals": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get list of rentals",
                "operationId": "a19748eac97c416a4a02a73fbcdb4700",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user"
                    },
                    {
                        "name": "car_id",
                        "in": "query",
                        "description": "Filter by car"
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter by start date"
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter by end date"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of rentals"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Create a new rental",
                "operationId": "7e700bb6638cc9e6a42aa112bf018ba6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "car_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "car_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rentals/{id}": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get rental details",
                "operationId": "7248fde8ecc49afa506485b780b4cd27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Update rental details",
                "operationId": "45134c777261098392b52d9f89a0311a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "car_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "car_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Delete a rental",
                "operationId": "f130db03982a2f6048b833c6cc977fcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rental deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}